requirements.txt
absl-py==0.9.0
adal==1.2.2
ansiwrap==0.8.4
applicationinsights==0.11.9
appnope==0.1.0
argcomplete==1.11.1
asn1crypto==1.3.0
astor==0.8.0
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
azure-common==1.1.25
azure-core==1.4.0
azure-graphrbac==0.61.1
azure-identity==1.2.0
azure-mgmt-authorization==0.60.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==8.0.1
azure-mgmt-storage==9.0.0
backcall==0.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
bleach==3.1.0
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
cachetools==3.1.1
catboost==0.23
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
configparser==3.7.4
contextlib2==0.6.0.post1
crcmod==1.7
cryptography==2.8
cycler==0.10.0
Cython==0.29.16
deap==1.3.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distro==1.5.0
docker==4.2.0
docopt==0.6.2
docutils==0.15.2
dotnetcore2==2.1.13
entrypoints==0.3
fasteners==0.15
fire==0.3.1
flake8==3.7.9
Flask==1.0.3
fusepy==3.0.1
future==0.18.2
gast==0.2.2
gcs-oauth2-boto-plugin==2.5
gensim==3.8.2
google-api-core==1.16.0
google-apitools==0.5.30
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
google-reauth==0.1.0
googleapis-common-protos==1.51.0
graphviz==0.13.2
grpcio==1.28.1
grpcio-tools==1.28.1
gsutil==4.49
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.17.2
idna==2.9
importlib-metadata==1.5.0
inference-schema==1.0.2
interpret-community==0.9.2
interpret-core==0.1.20
ipykernel==5.1.4
ipyparams==0.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json-logging-py==0.2
JsonForm==0.0.2
jsonpickle==1.4
jsonschema==3.2.0
JsonSir==0.0.2
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras2onnx==1.6.1
keyring==21.2.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
liac-arff==2.4.0
lightgbm==2.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==2.0.0
monotonic==1.5
msal==1.2.0
msal-extensions==0.1.3
msrest==0.6.13
msrestazure==0.6.3
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.4
ndg-httpsclient==0.5.1
nimbusml==1.7.0
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
onnx==1.6.0
onnxconverter-common==1.6.0
onnxmltools==1.4.1
opt-einsum==3.1.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
papermill==1.2.1
paramiko==2.7.1
parso==0.6.2
pathlib2==2.3.5
pathos==0.2.5
pathspec==0.8.0
patsy==0.5.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.4.10
pipreqsnb==0.1.3
pkginfo==1.5.0.1
plotly==4.7.1
pluggy==0.13.1
pmdarima==1.1.1
portalocker==1.7.0
pox==0.2.7
ppft==1.6.6.1
pprint==0.1
ppscore==0.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
py-cpuinfo==5.0.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
Python-EasyConfig==0.1.7
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2019.3
pyu2f==0.1.4
PyYAML==5.3.1
pyzmq==18.1.1
readme-renderer==26.0
remotely==0.0.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
Resource==0.2.1
retry-decorator==1.1.1
retrying==1.3.3
rope==0.17.0
rsa==4.0
ruamel.yaml==0.15.89
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
shap==0.34.0
simplekml==1.3.5
six==1.14.0
skl2onnx==1.4.9
sklearn-deap2==0.2.2
sklearn-pandas==1.7.0
smart-open==1.9.0
snowballstemmer==2.0.0
SocksiPy-branch==1.1
statsmodels==0.10.2
stopit==1.1.2
tabulate==0.8.7
tenacity==6.1.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-docs==0.0.0
tensorflow-estimator==2.1.0
tensorflow-serving-api==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textwrap3==0.9.2
toml==0.10.0
tornado==6.0.4
TPOT==0.11.1
tqdm==4.44.1
traitlets==4.3.3
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
ujson==1.35
Unidecode==1.1.1
update-checker==0.16
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
yapf==0.30.0
yarg==0.1.9
zipp==2.2.0
